Split out the main_full() demo from main().  Still need to add in the main_blinky() demo.
