+++ FreeRTOS/Demo/CORTEX_MPS2_QEMU_IAR/RTOSDemo.ewp
@@ -1222,4 +1222,7 @@
     <file>
         <name>$PROJ_DIR$\main.c</name>
     </file>
+    <file>
+        <name>$PROJ_DIR$\main_full.c</name>
+    </file>
 </project>
