+++ FreeRTOS/Demo/CORTEX_MPS2_QEMU_IAR/RTOSDemo.ewt
@@ -1348,4 +1348,7 @@
     <file>
         <name>$PROJ_DIR$\main.c</name>
     </file>
+    <file>
+        <name>$PROJ_DIR$\main_full.c</name>
+    </file>
 </project>
